requirements.txt
pyserial
pyusb
stlink
pystlink
